ACS-8541: Add userId in Questions and Answer API requests
